dom/worklet/tests/test_paintWorklet.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for PaintWorklet</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="common.js"></script>
</head>
<body>

<script type="application/javascript">

function configureTest() {
  function consoleListener() {
    SpecialPowers.addObserver(this, "console-api-log-event");
  }

  consoleListener.prototype  = {
    observe: function(aSubject, aTopic, aData) {
      if (aTopic == "console-api-log-event") {
        var obj = aSubject.wrappedJSObject;
        if (obj.arguments[0] == "So far so good") {
          ok(true, "Message received \\o/");

          SpecialPowers.removeObserver(this, "console-api-log-event");
          SimpleTest.finish();
          return;
        }
      }
    }
  }

  var cl = new consoleListener();

  return SpecialPowers.pushPrefEnv(
    {"set": [["dom.paintWorklet.enabled", true],
             ["dom.worklet.enabled", true]]});
}

// This function is called into an iframe.
function runTestInIframe() {
  paintWorklet.addModule("worklet_paintWorklet.js")
}
</script>
</body>
</html>
